build(gradle)!: Move script definitions to `script-utils`
Modules that contain custom scripts, for example as resources, must
depend on a module with the script definitions for IDE auto-completion
to work. Definitions and scripts cannot reside in the same module for that
purpose.

After this change, do the following in IntelliJ IDEA:
- Gradle sync
- Go to global settings
- Search the "Kotlin Scripting" section
- Expand "Custom Definitions" at the bottom
- Click on "Search Definitions"

That should make auto-completion for custom scripts work.

BREAKING CHANGE: Programmatic users of script definitions have to adapt to
the new artifact names.
