+++ evaluator/src/main/kotlin/Evaluator.kt
@@ -35,6 +35,7 @@ import org.ossreviewtoolkit.model.utils.DefaultResolutionProvider
 import org.ossreviewtoolkit.model.utils.ResolutionProvider
 import org.ossreviewtoolkit.model.utils.createLicenseInfoResolver
 import org.ossreviewtoolkit.utils.ort.Environment
+import org.ossreviewtoolkit.utils.script.RulesScriptTemplate
 import org.ossreviewtoolkit.utils.script.ScriptRunner
 
 class Evaluator(
